Attempt to fix duplicate assignments created
Not sure what's going on, doesn't seem like a code bug but due to the
nature of how App Engine works, it wants the apps to be stateless.
But I was storing the last run time in memory which is reset if you
start a new instance of the app however. So it might be that a new
instance was being created. Fix was to save last run time in the DB
and read it from there.