Delete assignments when task edited
This deletes any todo task assignments if the parent task is edited.
The field that don't show this behavior on being edited are
name and description of the task as they don't affect already scheduled
assignments in any way. The new name and description will be reflected
when client fetches again.