Attempt to fix duplicate assignments
From the research that was done, it seems that the issue is
not from multiple instances of the app but the same instance.

In the logs, https://console.cloud.google.com/logs/query;query=%0A;timeRange=2023-04-04T18:00:00.155Z%2F2023-04-04T18:05:16.155Z;cursorTimestamp=2023-04-04T18:04:36.040349Z?serviceId=default&versionId=20230309t023352&project=chores-326817
there are 2 statements that new assignments were added
and they both seem to be coming from the same version_id of
the project.

```
{
  "textPayload": "INFO: Adding 1 new assignments",
  "insertId": "...",
  "resource": {
    "type": "gae_app",
    "labels": {
      "project_id": "chores-326817",
      "zone": "us-east1-1",
      "module_id": "default",
      "version_id": "20230309t023352"
    }
  },
  "timestamp": "2023-04-04T18:04:34.330309Z",
  "labels": {
    "clone_id": "..."
  },
  "logName": "projects/chores-326817/logs/stderr",
  "receiveTimestamp": "2023-04-04T18:04:34.524467446Z"
}
```

```
{
  "textPayload": "INFO: Adding 1 new assignments",
  "insertId": "...",
  "resource": {
    "type": "gae_app",
    "labels": {
      "zone": "us-east1-1",
      "project_id": "chores-326817",
      "version_id": "20230309t023352",
      "module_id": "default"
    }
  },
  "timestamp": "2023-04-04T18:04:34.401428Z",
  "labels": {
    "clone_id": "..."
  },
  "logName": "projects/chores-326817/logs/stderr",
  "receiveTimestamp": "2023-04-04T18:04:34.524467446Z"
}
```

The version_ids are the same which probably means that it's
the same app. Which I'm not sure how it's possible because
the scheduler is kicking in only once. But if that is happening
this change will add more log statements that might confirm
that that is the issue.