Fix sync result member associations
Was returning only the member house association rows
which had the requesting member in them. But on client
side we need to be able to see which other members belong
to the same house so that we can assign tasks to them.

Fixed the response to send all member house associations
for all houses the requesting member belongs in.