Fix concurrent assignment update
Noticed an issue where if a client had marked an assignment as done
after another client had marked it as done, the server won't update
the status for second client and would not return it as an updated
assignemnt id. The clients were then not able to delete that assignment
from local storage and also just kept it around thinking that the update
didn't take place for some reason.

This fixes that by returning everything that was requested in the response.