Put google.json in resources
Looks like the file was added by GH actions into the repo
but it doesn't get bundled with the app when deploying to
Google Cloud App Engine.

Was seeing the file not found error for google.json.

Moved the file to resources which should get bundled with the app
and modified the action to write the file there.